Refactor VoiceRecorder component and remove VoiceRecorderSimple
